# backend/src/main/java/com/pliego/modules/sales/api/CheckoutResponse.java
package com.pliego.modules.sales.api;

public record CheckoutResponse(String orderId, String orderState, String paymentState, String total,
        String paymentReference, com.pliego.modules.sales.application.PostPurchaseModels.Fulfillment fulfillment,
        String subtotal, String taxRate, String taxAmount, String shippingAmount) { }