# frontend/openapi/openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "PLIEGO API",
    "description": "API de PLIEGO v1: autenticación, clientes, catálogo, inventario, carrito, pago y pedidos.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18445",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/checkout": {
      "post": {
        "tags": [
          "Checkout"
        ],
        "summary": "Finalizar compra simulada",
        "description": "No realiza cobros reales. Idempotency-Key identifica el intento; repetirlo devuelve el resultado original. Resuelve una respuesta perdida mediante el intento exacto.",
        "operationId": "checkout",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Pedido creado, con pago aprobado o rechazado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerJwt": []
          }
        ]
      }
    },
    "/api/v1/checkout/attempts/{key}/resolve": {
      "post": {
        "tags": [
          "Checkout"
        ],
        "summary": "Resolver un intento de compra",
        "description": "PENDING mantiene el resultado sin resolver. CREATED identifica el pedido exacto. NOT_CREATED bloquea permanentemente el intento antes de confirmar ausencia, incluso si la petición original llega tarde.",
        "operationId": "resolve",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Resultado del intento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutAttemptResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerJwt": []
          }
        ]
      }
    },
...
    "/api/v1/admin/orders/{orderId}/pickup/collect": {
      "post": {
        "tags": [
          "Administration — Orders"
        ],
        "summary": "Confirmar retiro del pedido",
        "description": "Comando ADMIN con referencia de retiro. Marca el pedido entregado y registra historial; repetirlo no duplica la entrega. La hora estimada no bloquea una entrega preparada antes.",
        "operationId": "collect",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "exclusiveMinimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerJwt": []
          }
        ]
      }
    },
...
    "/api/v1/orders": {
      "get": {
        "tags": [
          "Pedidos"
        ],
        "summary": "Consultar pedidos propios",
        "description": "Ordenados por fecha de creación y número de pedido descendentes.",
        "operationId": "list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 50,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Página de pedidos con resúmenes históricos de compra, entrega y factura",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseCustomerOrderSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerJwt": []
          }
        ]
      }
    },
...
    "/api/v1/admin/orders": {
      "get": {
        "tags": [
          "Pedidos administrativos"
        ],
        "summary": "Buscar pedidos",
        "description": "Filtros y orden definidos por fn_admin_orders.",
        "operationId": "search_1",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateFrom",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "dateTo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "customerId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "exclusiveMinimum": 0
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20,
              "maximum": 50,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Página de pedidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseAdminOrderSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerJwt": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetail": {
        "type": "object",
        "description": "Detalle de problema RFC 9457 con campos de correlación de PLIEGO.",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "example": "urn:pliego:problem:VALIDATION_ERROR"
          },
          "title": {
            "type": "string",
            "example": "Datos inválidos"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "Revisa los datos enviados e intenta nuevamente."
          },
          "instance": {
            "type": "string",
            "format": "uri",
            "example": "/api/v1/auth/register"
          },
          "code": {
            "type": "string",
            "description": "Código estable legible por máquina. Los errores de dominio usan el SQLSTATE canónico de PostgreSQL (Pxxxx); los errores propios de la capa REST usan un código simbólico (por ejemplo VALIDATION_ERROR o INVALID_CARD_NUMBER).",
            "example": "P3002"
          },
          "traceId": {
            "type": "string",
            "example": "8f2ab598-3e33-40b1-a5c7-15f676a84f9a"
          }
        }
      },
...
      "CheckoutRequest": {
        "type": "object",
        "properties": {
          "addressId": {
            "type": "string",
            "pattern": "^[1-9][0-9]{0,18}$"
          },
          "paymentMethod": {
            "type": "string",
            "pattern": "CARD|TRANSFER"
          },
          "simulationOutcome": {
            "type": "string"
          },
          "cardNumber": {
            "type": "string",
            "description": "Se valida localmente y nunca se persiste ni se envía a PostgreSQL.",
            "writeOnly": true
          },
          "expectedCartId": {
            "type": "string",
            "pattern": "^[1-9][0-9]{0,18}$"
          },
          "fulfillmentMethod": {
            "type": "string",
            "default": "HOME_DELIVERY",
            "enum": [
              "HOME_DELIVERY",
              "STORE_PICKUP"
            ],
            "pattern": "HOME_DELIVERY|STORE_PICKUP"
          },
          "pickupLocationId": {
            "type": "string",
            "pattern": "^[1-9][0-9]{0,18}$"
          }
        },
        "required": [
          "paymentMethod",
          "simulationOutcome"
        ]
      },
      "CheckoutResponse": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string"
          },
          "orderState": {
            "type": "string"
          },
          "paymentState": {
            "type": "string"
          },
          "total": {
            "type": "string"
          },
          "paymentReference": {
            "type": "string"
          },
          "fulfillment": {
            "$ref": "#/components/schemas/Fulfillment"
          },
          "subtotal": {
            "type": "string"
          },
          "taxRate": {
            "type": "string"
          },
          "taxAmount": {
            "type": "string"
          },
          "shippingAmount": {
            "type": "string"
          }
        }
      },
      "Fulfillment": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string"
          },
          "pickup": {
            "$ref": "#/components/schemas/Pickup"
          },
          "state": {
            "type": "string"
          },
          "collectedAt": {
            "type": "string"
          }
        }
      },
      "OpeningHours": {
        "type": "object",
        "properties": {
          "opensAt": {
            "type": "string"
          },
          "closesAt": {
            "type": "string"
          }
        }
      },
      "Pickup": {
        "type": "object",
        "properties": {
          "location": {
            "$ref": "#/components/schemas/PickupLocation"
          },
          "estimatedAt": {
            "type": "string"
          },
          "readyAt": {
            "type": "string"
          },
          "preparationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "pickupCode": {
            "type": "string"
          }
        }
      },
      "PickupLocation": {
        "type": "object",
        "description": "Punto de retiro con dirección y coordenadas estándar, independiente del proveedor de mapas.",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "province": {
            "type": "string"
          },
          "countryCode": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "timezone": {
            "type": "string"
          },
          "openingHours": {
            "$ref": "#/components/schemas/OpeningHours"
          },
          "active": {
            "type": "boolean"
          },
          "preparationMinutes": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CheckoutAttemptResponse": {
        "type": "object",
        "properties": {
          "state": {
            "type": "string"
          },
          "order": {
            "$ref": "#/components/schemas/CheckoutResponse"
          }
        }
      },
...
      "CollectRequest": {
        "type": "object",
        "properties": {
          "pickupCode": {
            "type": "string",
            "pattern": "P-[2-9A-HJ-NP-Z]{6,8}"
          }
        },
        "required": [
          "pickupCode"
        ]
      },
...
      "CustomerOrderSummary": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "orderState": {
            "type": "string"
          },
          "total": {
            "type": "string"
          },
          "paymentState": {
            "type": "string"
          },
          "purchaseState": {
            "type": "string"
          },
          "fulfillmentMethod": {
            "type": "string"
          },
          "shipmentState": {
            "type": "string"
          },
          "estimatedDeliveryFrom": {
            "type": "string"
          },
          "estimatedDeliveryTo": {
            "type": "string"
          },
          "itemCount": {
            "type": "integer",
            "format": "int64"
          },
          "unitCount": {
            "type": "integer",
            "format": "int64"
          },
          "itemSummary": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemSummary"
            }
          },
          "invoiceState": {
            "type": "string"
          },
          "invoicePdfAvailable": {
            "type": "boolean"
          },
          "invoiceXmlAvailable": {
            "type": "boolean"
          }
        }
      },
      "ItemSummary": {
        "type": "object",
        "properties": {
          "orderItemId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "format": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PageResponseCustomerOrderSummary": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerOrderSummary"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "string"
          }
        }
      },
...
      "AdminOrderSummary": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "customerName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "orderState": {
            "type": "string"
          },
          "total": {
            "type": "string"
          },
          "paymentState": {
            "type": "string"
          }
        }
      },
      "PageResponseAdminOrderSummary": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminOrderSummary"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerJwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "authSessionCookie": {
        "type": "apiKey",
        "name": "pliego_session",
        "in": "cookie"
      }
    }
  }
}